# Revert "Improve cached_op performance for static mode"