--- a/src/imperative/imperative_utils.h
+++ b/src/imperative/imperative_utils.h
@@ -595,21 +595,23 @@ inline bool CheckAndInferShape(nnvm::Graph* p_g, mxnet::ShapeVector&& shapes,
     *contain_unknown = false;
   }
   nnvm::Graph& g = *p_g;
-  if (g.attrs.count("shape")) {
+  if (use_inputs) {
+    if (g.attrs.count("shape_inputs") &&
+        g.GetAttr<mxnet::ShapeVector>("shape_inputs") == shapes) return true;
+  } else if (g.attrs.count("shape")) {
     const auto& prev_shapes = g.GetAttr<mxnet::ShapeVector>("shape");
-    if (prev_shapes.size() == shapes.size()) {
-      bool match = true;
-      for (size_t i = 0; i < shapes.size(); ++i) {
-        if (i == entry_range.first) {
-          i = entry_range.second;
-          if (i >= shapes.size()) break;
-        }
-        if (shapes[i] == prev_shapes[i]) continue;
-        match = false;
-        break;
+    CHECK_EQ(prev_shapes.size(), shapes.size());
+    bool match = true;
+    for (size_t i = 0; i < shapes.size(); ++i) {
+      if (i == entry_range.first) {
+        i = entry_range.second;
+        if (i >= shapes.size()) break;
       }
-      if (match) return true;
+      if (shapes[i] == prev_shapes[i]) continue;
+      match = false;
+      break;
     }
+    if (match) return true;
   }
   g.attrs.erase("shape");
   g.attrs.erase("shape_inputs");