@eric-haibin-lin 2019-05-04
@ZhennanQin could you look into why this PR introduces the regression?